refactor: move private imp to detail
