+++ common/autoware_boundary_departure_checker/include/autoware/boundary_departure_checker/detail/footprints_generator.hpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__BOUNDARY_DEPARTURE_CHECKER__FOOTPRINTS_GENERATOR_HPP_
-#define AUTOWARE__BOUNDARY_DEPARTURE_CHECKER__FOOTPRINTS_GENERATOR_HPP_
+#ifndef AUTOWARE__BOUNDARY_DEPARTURE_CHECKER__DETAIL__FOOTPRINTS_GENERATOR_HPP_
+#define AUTOWARE__BOUNDARY_DEPARTURE_CHECKER__DETAIL__FOOTPRINTS_GENERATOR_HPP_
 
-#include <autoware/boundary_departure_checker/side_struct.hpp>
+#include <autoware/boundary_departure_checker/detail/side_struct.hpp>
 #include <autoware_utils_geometry/boost_geometry.hpp>
 #include <autoware_utils_geometry/geometry.hpp>
 #include <autoware_vehicle_info_utils/vehicle_info.hpp>
@@ -70,4 +70,4 @@ FootprintMargin calc_margin_from_covariance(
 
 }  // namespace autoware::boundary_departure_checker::footprints
 
-#endif  // AUTOWARE__BOUNDARY_DEPARTURE_CHECKER__FOOTPRINTS_GENERATOR_HPP_
+#endif  // AUTOWARE__BOUNDARY_DEPARTURE_CHECKER__DETAIL__FOOTPRINTS_GENERATOR_HPP_
