+++ common/autoware_boundary_departure_checker/src/detail/conversion.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/boundary_departure_checker/conversion.hpp"
+#include "autoware/boundary_departure_checker/detail/conversion.hpp"
 
 namespace autoware::boundary_departure_checker::utils
 {
