+++ common/autoware_boundary_departure_checker/src/detail/severity_evaluator.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/boundary_departure_checker/severity_evaluator.hpp"
+#include "autoware/boundary_departure_checker/detail/severity_evaluator.hpp"
 
 #include <autoware/motion_utils/distance/distance.hpp>
