+++ common/autoware_boundary_departure_checker/test/test_uncrossable_boundary_checker.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/boundary_departure_checker/type_alias.hpp"
+#include "autoware/boundary_departure_checker/detail/type_alias.hpp"
 #include "autoware/boundary_departure_checker/uncrossable_boundary_checker.hpp"
 
 #include <autoware_vehicle_info_utils/vehicle_info_utils.hpp>
